+++ src/constants/selectOptions.js
@@ -23,6 +23,7 @@ export const OPTIONS = {
   ADMINISTRATIVE_NOTE: 'ADMINISTRATIVE_NOTE',
   CHECK_IN_NOTE: 'CHECK_IN_NOTE',
   CHECK_OUT_NOTE: 'CHECK_OUT_NOTE',
+  URL_RELATIONSHIP: 'URL_RELATIONSHIP',
 };
 
 export const PARAMETERS_KEYS = {
@@ -32,6 +33,15 @@ export const PARAMETERS_KEYS = {
 
 export const OPTIONS_WITH_ADDITIONAL_PARAMETERS = [
   OPTIONS.ITEM_NOTE,
+];
+
+export const getItemsWithPlaceholder = (items) => [
+  {
+    value: '',
+    label: <FormattedMessage id="ui-bulk-edit.list.filters.recordIdentifier.placeholder" />,
+    disabled: true,
+  },
+  ...items,
 ];
 
 export const identifierOptions = {
@@ -194,6 +204,12 @@ export const getHoldingsOptions = (formatMessage, holdingsNotes = []) => [
     categoryName: formatMessage({ id: 'ui-bulk-edit.category.holdingsLocation' }),
   },
   ...holdingsNotes,
+  {
+    value: OPTIONS.URL_RELATIONSHIP,
+    label: formatMessage({ id: 'ui-bulk-edit.layer.options.holdings.urlRelationship' }),
+    disabled: false,
+    categoryName: formatMessage({ id: 'ui-bulk-edit.category.electronicAccess' }),
+  },
   {
     value: OPTIONS.SUPPRESS_FROM_DISCOVERY,
     label: formatMessage({ id: 'ui-bulk-edit.layer.options.holdings.suppress' }),